move db out of domain
